+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] 2026-02-27
+### Added
+- Option in `make_samplelists_from_manifest.pl` to include tumour type from a metadata file
+
+### Changed
+- Updated CI/CD template to use ref 0.4.2
+- Updated build to inherit from MAF 'latest'
 
 ## [0.5.2] 2025-02-20
 ### Added